requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
asyncpg==0.29.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==1.10.11
python-multipart==0.0.6
aiocache==0.11.1
python-dotenv==1.0.0
email-validator==2.0.0.post2